requirements.txt
fastmcp
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

#
fastmcp
pydantic
google-api-python-client
google-auth
uvicorn
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
requests